fix: warning and error types in own file
Place the type definition for the warning and error callbacks on their own files, this way noise is removed from converter.ts file
