+++ src/utils/getBat.spec.ts
@@ -2,8 +2,9 @@ import { describe, it } from 'node:test'
 import assert from 'node:assert'
 import { getBat } from './getBat.js'
 import { type BatteryData } from '@nordicsemiconductor/asset-tracker-cloud-docs'
-import { ValidationError, UndefinedLwM2MObjectWarning } from '../converter.js'
 import { Device_3_urn, parseURN } from '@nordicsemiconductor/lwm2m-types'
+import type { UndefinedLwM2MObjectWarning } from './UndefinedLwM2MObjectWarning.js'
+import type { ValidationError } from './ValidationError.js'
 
 void describe('getBat', () => {
 	void it(`should create the 'bat' object expected by 'nRF Asset Tracker Reported'`, () => {
