+++ src/utils/getCfg.ts
@@ -4,7 +4,8 @@ import {
 	type ConfigData,
 } from '@nordicsemiconductor/asset-tracker-cloud-docs/protocol'
 import { validateAgainstSchema } from './validateAgainstSchema.js'
-import { ValidationError, UndefinedLwM2MObjectWarning } from '../converter.js'
+import type { ValidationError } from './ValidationError.js'
+import { UndefinedLwM2MObjectWarning } from './UndefinedLwM2MObjectWarning.js'
 
 /**
  * Defines the result type of 'getCfg' method, which will be one of the following options:
