+++ src/utils/getEnv.spec.ts
@@ -2,8 +2,9 @@ import { describe, it } from 'node:test'
 import assert from 'node:assert'
 import { Humidity_3304_urn, type Temperature_3303 } from '../schemas/index.js'
 import { getEnv } from './getEnv.js'
-import { ValidationError, UndefinedLwM2MObjectWarning } from '../converter.js'
 import { parseURN } from '@nordicsemiconductor/lwm2m-types'
+import type { UndefinedLwM2MObjectWarning } from './UndefinedLwM2MObjectWarning.js'
+import type { ValidationError } from './ValidationError.js'
 
 void describe('getEnv', () => {
 	void it(`should create the 'env' object expected by 'nRF Asset Tracker Reported'`, () => {
