+++ src/utils/getGnss.spec.ts
@@ -2,8 +2,9 @@ import { describe, it } from 'node:test'
 import assert from 'node:assert'
 import { Location_6_urn, type Location_6 } from '../schemas/index.js'
 import { getGnss } from './getGnss.js'
-import { ValidationError, UndefinedLwM2MObjectWarning } from '../converter.js'
 import { parseURN } from '@nordicsemiconductor/lwm2m-types'
+import type { UndefinedLwM2MObjectWarning } from './UndefinedLwM2MObjectWarning.js'
+import type { ValidationError } from './ValidationError.js'
 
 void describe('getGnss', () => {
 	void it(`should create the 'gnss' object expected by 'nRF Asset Tracker Reported'`, () => {
